requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
mysqlclient==2.2.0
drf-yasg==1.21.7
python-decouple==3.8
django-filter==23.5
Pillow==10.1.0
gunicorn==22.0.0
celery==5.3.6
redis==5.0.4
requests==2.31.0 